CORE-147 - New Landing Page (#1224)
