+++ resources/views/layout.blade.php
@@ -41,7 +41,7 @@
             @endif
             <div class="nav_upper_container navbar-toggleable-md">
                 <div class="logo_container">
-                    <a href="{{ route("default") }}">
+                    <a href="{{ route("dashboard") }}">
                         {!! HTML::image("images/vatsim_uk_logo.png", "UK Logo") !!}
                     </a>
                 </div>
@@ -55,7 +55,7 @@
                 <div class="collapse navbar-collapse js-navbar-collapse" id="nav">
                     <ul class="nav navbar-nav navcustom">
                         <li class="dropdown dropdown-large">
-                            {!! link_to_route("default", "Home") !!}
+                            {!! link_to_route("dashboard", "Home") !!}
                         </li>
                     </ul>
                                                        
@@ -82,7 +82,7 @@
                                         <li>{{ HTML::link('https://vatsim.uk/', 'VATSIM UK Homepage', array("target"=>"_blank")) }}</li>
                                         <li>{{ HTML::link('https://cts.vatsim.uk/', 'Central Training System', array("target"=>"_blank")) }}</li>
                                         <li>{{ HTML::link('http://www.nats-uk.ead-it.com/public/index.php%3Foption=com_content&task=blogcategory&id=6&Itemid=13.html', 'UK Charts', array("target"=>"_blank")) }}</li>
-                                        <li>{!! link_to_route("airports", "UK Airfields") !!}</li>
+                                        <li>{!! link_to_route("site.airports", "UK Airfields") !!}</li>
                                         <li class="divider"></li>
                                         <li class="dropdown-header">Network Statistics</li>
                                         <li>{!! link_to_route("networkdata.dashboard", "My Statistics") !!}</li>
