+++ resources/views/site/home.blade.php
@@ -1,0 +1,8 @@
+<html>
+<head>
+    <title>Placeholder</title>
+</head>
+<body>
+<h1>Placeholder</h1>
+</body>
+</html>
