+++ tests/Unit/MshipAuthenticationTest.php
@@ -65,7 +65,7 @@ class MshipAuthenticationTest extends TestCase
     {
         // In this test, we assert that we're getting a redirect to the login page if no auth at all
         $expectedStatus = 302;
-        $expectedRedirectRegExp = '#\h*<title>Redirecting to '.route('default').'<\/title>\h*#';
+        $expectedRedirectRegExp = '#\h*<title>Redirecting to '.route('dashboard').'<\/title>\h*#';
         $expectedObjectType = 'Illuminate\Http\RedirectResponse';
 
         $result = $this->authenticationInstance->getLogin();
