+++ webpack.mix.js
@@ -27,5 +27,8 @@ mix.less('resources/assets/less/admin.less', 'public/css/admin-all.css');
 mix.sass('resources/assets/sass/app.scss', 'public/css/app-all.css');
 mix.js('resources/assets/js/app.js', 'public/js/app-all.js');
 
+// homepage
+mix.sass('resources/assets/sass/home.scss', 'public/css/home.css');
+
 // general
 mix.version();
